pyproject.toml
[tool.poetry]
name = "seedlot-selection-tool"
version = "1.0.0"
description = "A web-based spatial analysis and mapping tool that allows managers to find appropriate seedlots for a given planting site, given information about species, transfer limits, and historic and future climate projections across the landscape."
authors = ["Conservation Biology Institute"]
license = "BSD-3-Clause"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
trefoil = { git = "https://github.com/consbio/trefoil.git", branch = "main" }
rasterio = "^1.2a1"
django-celery-results = "^2.0.0"
social-auth-app-django = "^5.0.0"
psycopg2-binary = "^2.9.10"
django-webpack-loader = "^1.1.0"
celery = "^5.2.3"
pyproj = "^3.6.1"
sentry-sdk = "^0.19.5"
enmerkar = "^0.7.1"
ncdjango = { path = "./ncdjango" }
aiohttp = "^3.12.3"
Django = "^5.2"
djangorestframework = "~3.14"
GDAL = { version = "*", optional = true }
geopy = "^1.23.0"
mercantile = "^1.1.6"
netCDF4 = "^1.5.5"
progress = "^1.5"
python-pptx = "^0.6.21"
WeasyPrint = "^52.2"
django-waffle = "^2.2"
django-url-filter = "^0.3.15"

[poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
